// src/app/page/master/detail-jenis-produk/detail-jenis-produk.interface.ts
import {EventEmitter} from '@angular/core';

export interface DetailJenisProduk {
	namaDetailJenisProduk?: string;//
	 kdJenisProduk?: string;//
     persenHargaCito?: number;//
	 kdAccount?: number;    
     isRegistrasiAset?: number;
     // kdDepartemen?: string;

	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
}